Added animations to the page
