+++ src/app/globals.css
@@ -9,3 +9,4 @@
   color: transparent;
   fallback: #000;
 }
+
