+++ src/components/ContactForm.tsx
@@ -1,10 +1,12 @@
-'use client'
-
 import React from 'react'
 import { useForm, ValidationError } from '@formspree/react'
 import { RiMailSendFill } from 'react-icons/ri'
 
-export default function ContactForm() {
+type ContactProps = {
+  addToRefs: any
+}
+
+export default function ContactForm({ addToRefs }: ContactProps) {
   const [state, handleSubmit] = useForm('mgebqlnb')
 
   if (state.succeeded) {
@@ -18,7 +20,10 @@ export default function ContactForm() {
   }
 
   return (
-    <main className='3xl:py-[5rem] flex flex-col justify-center items-center gap-10 mx-auto w-full '>
+    <main
+      ref={addToRefs}
+      className='3xl:py-[5rem] xl:px-0 md:px-10 md:py-[3rem] sm:px-5 sm:py-[2rem] flex flex-col justify-center items-center gap-10 mx-auto w-full '
+    >
       <h1 className='text-gradient xl:text-5xl md:text-4xl sm:text-3xl capitalize'>
         Contact
       </h1>
