+++ src/components/Footer.tsx
@@ -1,5 +1,5 @@
-import Link from 'next/link'
 import React from 'react'
+import Link from 'next/link'
 import { BsGithub, BsInstagram, BsLinkedin, BsTwitter } from 'react-icons/bs'
 
 const navLinks = [
